+++ apps/fluent-tester/src/FluentTester/TestComponents/MenuButton/MenuButtonTestStyles.ts
@@ -1,0 +1,5 @@
+import { ViewStyle, TextStyle } from 'react-native';
+export const viewWrapperStyle: ViewStyle = { flexDirection: 'row', paddingVertical: 5 };
+export const columnStyle: ViewStyle = { flexDirection: 'column', paddingHorizontal: 5 };
+export const rowStyle: ViewStyle = { flexDirection: 'row' };
+export const textColor: TextStyle = { color: 'blue' };