+++ apps/fluent-tester/src/FluentTester/TestComponents/TabsExperimental/consts.ts
@@ -1,2 +1,2 @@
-export const HOMEPAGE_TABS_BUTTONEXPERIMENTAL = 'Homepage_Tabs_ButtonExperimental';
+export const HOMEPAGE_EXPERIMENTAL_TABS_BUTTON = 'Homepage_Tabs_ButtonExperimental';
 export const TABS_TESTPAGE = 'Tabs_TestPage';