+++ apps/fluent-tester/src/FluentTester/testPages.win32.ts
@@ -19,7 +19,8 @@ import { HOMEPAGE_THEME_BUTTON, ThemeTest } from './TestComponents/Theme';
 import { HOMEPAGE_TOKEN_BUTTON, TokenTest } from './TestComponents/Tokens';
 import { HOMEPAGE_ICON_BUTTON, IconTest } from './TestComponents/Icon';
 import { HOMEPAGE_TABS_BUTTON, TabsTest } from './TestComponents/Tabs';
-import { HOMEPAGE_TABS_BUTTONEXPERIMENTAL, TabsExperimentalTest } from './TestComponents/TabsExperimental';
+import { HOMEPAGE_EXPERIMENTAL_TABS_BUTTON, TabsExperimentalTest } from './TestComponents/TabsExperimental';
+import { HOMEPAGE_EXPERIMENTAL_TEXT_BUTTON, TextExperimentalTest } from './TestComponents/TextExperimental';
 import { TestDescription } from './TestComponents';
 
 export const tests: TestDescription[] = [
@@ -114,6 +115,11 @@ export const tests: TestDescription[] = [
     testPage: HOMEPAGE_TEXT_BUTTON,
   },
   {
+    name: 'Experimental Text Test',
+    component: TextExperimentalTest,
+    testPage: HOMEPAGE_EXPERIMENTAL_TEXT_BUTTON,
+  },
+  {
     name: 'Theme Test',
     component: ThemeTest,
     testPage: HOMEPAGE_THEME_BUTTON,
@@ -126,7 +132,7 @@ export const tests: TestDescription[] = [
   {
     name: 'Tabs Experimental Test',
     component: TabsExperimentalTest,
-    testPage: HOMEPAGE_TABS_BUTTONEXPERIMENTAL,
+    testPage: HOMEPAGE_EXPERIMENTAL_TABS_BUTTON,
   },
   {
     name: 'Tokens Test',