+++ packages/components/ContextualMenu/src/SubmenuItem.settings.ts
@@ -7,7 +7,7 @@ export const settings: IComposeSettings<SubmenuItemType> = [
       backgroundColor: 'menuBackground',
       color: 'menuItemText',
       borderColor: 'transparent',
-      borderWidth: 1,
+      borderWidth: 2,
     },
     root: {
       accessible: true,
@@ -60,7 +60,7 @@ export const settings: IComposeSettings<SubmenuItemType> = [
       },
       pressed: {
         tokens: {
-          backgroundColor: 'menuItemBackgroundHovered',
+          backgroundColor: 'menuItemBackgroundPressed',
           color: 'menuItemTextHovered',
         },
       },
@@ -74,20 +74,12 @@ export const settings: IComposeSettings<SubmenuItemType> = [
         tokens: {
           color: 'menuItemTextHovered',
           backgroundColor: 'menuItemBackgroundHovered',
+          borderColor: 'focusBorder'
         },
         _overrides: {
-          disabled: {
-            tokens: {
-              borderColor: 'focusBorder',
-            },
-          },
           hovered: {
-            _overrides: {
-              disabled: {
-                tokens: {
-                  borderColor: 'transparent',
-                },
-              },
+            tokens: {
+              borderColor: 'transparent'
             },
           },
         },