+++ packages/utils/tokens/src/color-tokens.ts
@@ -1,4 +1,4 @@
-import { ITheme, IThemeColorDefinition } from '@uifabricshared/theming-ramp';
+import { Theme, ThemeColorDefinition } from '@fluentui-react-native/theme-types';
 import { IOperationSet } from '@uifabricshared/foundation-tokens';
 import { ColorValue } from 'react-native';
 
@@ -6,17 +6,17 @@ export interface IForegroundColorTokens {
   color?: ColorValue;
 }
 
-export const getPaletteFromTheme = (theme: ITheme): IThemeColorDefinition => {
+export const getPaletteFromTheme = (theme: Theme): ThemeColorDefinition => {
   return theme.colors;
 };
 
-export const foregroundColorTokens: IOperationSet<IForegroundColorTokens, ITheme> = [{ source: 'color', lookup: getPaletteFromTheme }];
+export const foregroundColorTokens: IOperationSet<IForegroundColorTokens, Theme> = [{ source: 'color', lookup: getPaletteFromTheme }];
 
 export interface IBackgroundColorTokens {
   backgroundColor?: ColorValue;
 }
 
-export const backgroundColorTokens: IOperationSet<IBackgroundColorTokens, ITheme> = [
+export const backgroundColorTokens: IOperationSet<IBackgroundColorTokens, Theme> = [
   { source: 'backgroundColor', lookup: getPaletteFromTheme },
 ];