+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Fri, 13 Aug 2021 16:12:26 GMT and should not be manually modified.
+This log was last generated on Mon, 23 Aug 2021 23:35:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.13
+
+Mon, 23 Aug 2021 23:35:47 GMT
+
+### Patches
+
+- Fixed icon color for SVG (v.kozlova13@gmail.com)
+
+## 0.8.12
+
+Thu, 19 Aug 2021 18:31:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.10.10 (ruaraki@microsoft.com)
+
+## 0.8.11
+
+Wed, 18 Aug 2021 23:00:57 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.10.9 (ruaraki@microsoft.com)
+
+## 0.8.10
+
+Wed, 18 Aug 2021 20:17:46 GMT
+
+### Patches
+
+- Fix iconColor not applying for svg icons (ruaraki@microsoft.com)
 
 ## 0.8.9