+++ packages/framework/framework/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Fri, 13 Aug 2021 16:12:26 GMT and should not be manually modified.
+This log was last generated on Mon, 23 Aug 2021 23:35:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.32
+
+Mon, 23 Aug 2021 23:35:47 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tokens to v0.9.20 (dake.3601@gmail.com)
+
+## 0.5.31
+
+Thu, 19 Aug 2021 18:31:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tokens to v0.9.19 (ruaraki@microsoft.com)
+
+## 0.5.30
+
+Wed, 18 Aug 2021 23:00:57 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.7.12 (ruaraki@microsoft.com)
 
 ## 0.5.29